+++ client/_constants/apiEndpoints.ts
@@ -27,12 +27,12 @@ const endPoints = {
   frostlakeCreateAnalyticsEndpoint: `${apiHost}/api/frostlake/createanalytics`,
   snowlakeCreateTxEndpoint: `${apiHost}/api/snowlake/createtx`,
   swapstreamCreateTxEndpoint: `${apiHost}/api/swapstream/createtx`,
-  vuelockCreateVaultEndpoint: `${apiHost}/api/vuelock/createvault`,
-  vuelockGetMyVaultsEndpoint: `${apiHost}/api/vuelock/getmyvaults`,
-  vuelockViewVaultEndpoint: `${apiHost}/api/vuelock/viewvault`,
-  vuelockDeleteVaultEndpoint: `${apiHost}/api/vuelock/deletevault`,
-  vuelockCreateSecretEndpoint: `${apiHost}/api/vuelock/createsecret`,
-  vuelockDeleteSecretEndpoint: `${apiHost}/api/vuelock/deletesecret`,
+  hyperedgeCreateDbEndpoint: `${apiHost}/api/hyperedge/createdb`,
+  hyperedgeGetMyDbsEndpoint: `${apiHost}/api/hyperedge/getmydbs`,
+  hyperedgeViewDbEndpoint: `${apiHost}/api/hyperedge/viewdb`,
+  hyperedgeDeleteDbEndpoint: `${apiHost}/api/hyperedge/deletedb`,
+  hyperedgeCreateKvEndpoint: `${apiHost}/api/hyperedge/createkv`,
+  hyperedgeDeleteKvEndpoint: `${apiHost}/api/hyperedge/deletekv`,
   swapstreamTokenConfigEndpoint: `${apiHost}/api/swapstream/getswapstreamtokenconfig`,
 }