+++ src/api/apps/apps.module.ts
@@ -4,9 +4,9 @@ import { FrostlakeModule } from "./frostlake/frostlake.module"
 import { SwapstreamModule } from "./swapstream/swapstream.module"
 import { DwalletModule } from "./dwallet/dwallet.module"
 import { SnowlakeModule } from "./snowlake/snowlake.module"
-import { VuelockModule } from "./vuelock/vuelock.module"
+import { HyperedgeModule } from "./hyperedge/hyperedge.module"
 
 @Module({
-  imports: [AirlakeModule, FrostlakeModule, SwapstreamModule, DwalletModule, SnowlakeModule, VuelockModule]
+  imports: [AirlakeModule, FrostlakeModule, SwapstreamModule, DwalletModule, SnowlakeModule, HyperedgeModule]
 })
 export class AppsModule { }