+++ src/api/apps/hyperedge/dto/create-db.dto.ts
@@ -1,6 +1,6 @@
 import { IsNotEmpty } from "class-validator"
 
-export class CreateVaultDto {
+export class CreateDbDto {
   @IsNotEmpty()
   name: string
 }